requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-multipart==0.0.9
faiss-cpu==1.13.2
sentence-transformers==2.7.0
ollama==0.2.0
PyPDF2==3.0.1
python-docx==1.1.2
python-dotenv==1.0.1
requests==2.32.3
tqdm==4.66.4
markdown==3.6
rank-bm25==0.2.2

# DeepSeek API client
openai==1.30.2
